Collections: Updates associated tests
